useIsTrackAvailable condition corrected, and removed unnecessary files
